requirements.txt
pypdf2
pdfplumber
spacy
sentence-transformers
scikit-learn
hdbscan
umap-learn
streamlit
pandas
numpy
